Split svc.go into separate files
